Work on optimizing some component organization and functionality. Moved remove student action to option menu dropdown.
