+++ app/controllers/utils/ioEmit.js
@@ -1,6 +1,7 @@
-module.exports = (req, io, action, payload) => {
+module.exports = (req, io, action, payload, force = false) => {
     io.emit( action, {
         payload,
-        from: req.userSocket && req.userSocket.id
+        from: req.userSocket && req.userSocket.id,
+        force
     } );
 }
