+++ client/src/pages/Dashboard/views/Students/components/StudentList.js
@@ -26,19 +26,17 @@ const StudentList = ( { sort, groupBy, search } ) => {
             {
                 groupedStudents[0].group.key === "default"
 
-                ? groupedStudents[0].entries.filter( studentFilter ).sort( studentSort ).map(student => {
-                    return (
-                        <Columns.Column key={student._id} {...sizes} className="has-filled-content">
-                            <StudentCard student={student} />
-                        </Columns.Column>
-                    )
-                })
+                ? groupedStudents[0].entries.filter( studentFilter ).sort( studentSort ).map(student => (
+                    <Columns.Column key={student._id} {...sizes} className="has-filled-content">
+                        <StudentCard student={student} />
+                    </Columns.Column>
+                ))
 
                 : groupedStudents.map( ({ group, entries }) => (
-                        <Columns.Column key={group.key} {...sizes}>
-                            <Card className={`is-${group.color} is-light w-100 has-shadow`}><Card.Content>{group.label}</Card.Content></Card>
-                            {entries.filter( studentFilter ).sort( studentSort ).map( student => <StudentCard key={student._id} student={student} className="mt-4" /> )}
-                        </Columns.Column>
+                    <Columns.Column key={group.key} {...sizes}>
+                        <Card className={`is-${group.color} is-light w-100 has-shadow`}><Card.Content>{group.label}</Card.Content></Card>
+                        {entries.filter( studentFilter ).sort( studentSort ).map( student => <StudentCard key={student._id} student={student} className="mt-4" /> )}
+                    </Columns.Column>
                 ) )
             }
         </Columns>
