+++ client/src/pages/Dashboard/views/Students/index.js
@@ -5,7 +5,6 @@ import { Section } from "react-bulma-components";
 import { useTopbarConfig } from "pages/Dashboard/components/Topbar";
 
 import StudentList from "./components/StudentList";
-import { ModalProvider } from "components/Modal";
 import StudentModal from "../../components/StudentModal"
 import StudentListControls from "./components/StudentListControls";
 
@@ -18,11 +17,9 @@ const Students = () => {
 
     return (
         <Section>
-            <ModalProvider>
-                <StudentListControls sort={[ sort, setSort ]} groupBy={[ groupBy, setGroupBy ]} search={[ search, setSearch ]} />
-                <StudentList sort={sort} groupBy={groupBy} search={search.toLowerCase()} />
-                <StudentModal />
-            </ModalProvider>
+            <StudentListControls sort={[ sort, setSort ]} groupBy={[ groupBy, setGroupBy ]} search={[ search, setSearch ]} />
+            <StudentList sort={sort} groupBy={groupBy} search={search.toLowerCase()} />
+            <StudentModal />
         </Section>
     );
